perf: use lightweight epoch fetch for historical epochs (~25x faster)
Historical epoch fetches used the SDK's getEpoch() which makes ~55 RPC
calls per epoch (per-gateway weight lookups, name resolution,
observations). Replace with fetchEpochLightweight() which reads the
Epoch PDA directly (1 RPC call) and builds the EpochData from the raw
on-chain data.

Extract shared fetchEpochLightweight into src/utils/epochFetch.ts,
used by both GlobalDataProvider (current epoch) and store/db.ts
(historical epochs with IndexedDB caching).
